+++ subscription/api/blocking/src/main/java/org/occurrent/subscription/api/blocking/SubscriptionModelCancelSubscription.java
@@ -1,9 +1,0 @@
-package org.occurrent.subscription.api.blocking;
-
-public interface SubscriptionModelCancelSubscription {
-    /**
-     * Cancel a subscription, this will remove the position from position storage (if used),
-     * and you cannot restart it from its current position again.
-     */
-    void cancelSubscription(String subscriptionId);
-}