+++ subscription/api/blocking/src/main/java/org/occurrent/subscription/api/blocking/SubscriptionModelLifeCycle.java
@@ -3,7 +3,7 @@ package org.occurrent.subscription.api.blocking;
 /**
  * Defines life-cycle methods for subscription models and subscriptions.
  */
-public interface SubscriptionModelLifeCycle extends SubscriptionModelCancelSubscription {
+public interface SubscriptionModelLifeCycle {
 
     /**
      * Temporary stop the subscription model so that none of its subscriptions will receive any events.
@@ -58,4 +58,18 @@ public interface SubscriptionModelLifeCycle extends SubscriptionModelCancelSubsc
      * @throws IllegalArgumentException If subscription is not running
      */
     void pauseSubscription(String subscriptionId);
+
+    /**
+     * Cancel a subscription, this will remove the position from position storage (if used),
+     * and you cannot restart it from its current position again.
+     */
+    void cancelSubscription(String subscriptionId);
+
+    /**
+     * Shutdown the subscription model and close all subscriptions (they can be resumed later if you start from a durable subscription position).
+     * A subscription model that is shutdown cannot be started again, since it closes resources such as database connections,
+     * thread pools etc.
+     */
+    default void shutdown() {
+    }
 }